refactor: Replace TradeEventListener with OrderEventListener to handle order cancellation and expiration events.
